+++ mrpt_graphslam_2d/include/mrpt_graphslam_2d/TNeighborAgentMapProps.h
@@ -12,7 +12,7 @@
 #include <ros/ros.h>
 #include <nav_msgs/OccupancyGrid.h>
 #include <nav_msgs/Path.h>
-#include <mrpt_bridge/map.h>
+#include <mrpt/ros1bridge/map.h>
 #include <mrpt_msgs/GraphSlamAgents.h>
 #include <mrpt/system/COutputLogger.h>
 #include <mrpt/maps/COccupancyGridMap2D.h>