+++ mrpt_graphslam_2d/include/mrpt_graphslam_2d/misc/common.h
@@ -1,11 +1,11 @@
 /* +---------------------------------------------------------------------------+
-	 |                     Mobile Robot Programming Toolkit (MRPT)               |
-	 |                          http://www.mrpt.org/                             |
-	 |                                                                           |
-	 | Copyright (c) 2005-2016, Individual contributors, see AUTHORS file        |
-	 | See: http://www.mrpt.org/Authors - All rights reserved.                   |
-	 | Released under BSD License. See details in http://www.mrpt.org/License    |
-	 +---------------------------------------------------------------------------+ */
+	 |                     Mobile Robot Programming Toolkit (MRPT) | |
+   http://www.mrpt.org/                             | | | | Copyright (c)
+   2005-2016, Individual contributors, see AUTHORS file        | | See:
+   http://www.mrpt.org/Authors - All rights reserved.                   | |
+   Released under BSD License. See details in http://www.mrpt.org/License    |
+	 +---------------------------------------------------------------------------+
+ */
 
 /**\file common.h
  *
@@ -17,20 +17,20 @@
 #include <mrpt/poses/CPose2D.h>
 #include <mrpt/slam/CGridMapAligner.h>
 #include <sstream>
-
-#include <mrpt/version.h>
-#if MRPT_VERSION>=0x199
 #include <mrpt/containers/stl_containers_utils.h>
-#else
-#include <mrpt/system/stl_containers_utils.h>
-#endif
 
-namespace mrpt { namespace graphslam { namespace detail {
-
+namespace mrpt
+{
+namespace graphslam
+{
+namespace detail
+{
 std::string getGridMapAlignmentResultsAsString(
-			const mrpt::poses::CPosePDF& pdf,
-			const mrpt::slam::CGridMapAligner::TReturnInfo& ret_info);
+	const mrpt::poses::CPosePDF& pdf,
+	const mrpt::slam::CGridMapAligner::TReturnInfo& ret_info);
 
 bool isEssentiallyZero(const mrpt::poses::CPose2D& p);
 
-} } } // end of namespaces
+}  // namespace detail
+}  // namespace graphslam
+}  // namespace mrpt